fix: adjust extraordinary
