+++ src/pages/prospect/components/TableExtraordinaryInstallment/index.tsx
@@ -1,8 +1,10 @@
 import { useEffect, useState } from "react";
-import { useMediaQuery } from "@inubekit/inubekit";
+import { useMediaQuery, useFlag } from "@inubekit/inubekit";
 
 import { IProspect } from "@services/prospects/types";
 import { IExtraordinaryInstallments } from "@services/prospect/types/extraordInaryInstallments";
+import { TextLabels } from "@components/modals/ExtraordinaryPaymentModal/config";
+import { removeExtraordinaryInstallment } from "./utils";
 
 import {
   headersTableExtraordinaryInstallment,
@@ -82,24 +84,45 @@ export const TableExtraordinaryInstallment = (
   } = props;
 
   const headers = headersTableExtraordinaryInstallment;
-  const [extraordinaryInstallments, setExtraordinaryInstallments] = useState<
-    TableExtraordinaryInstallmentProps[]
-  >([]);
-  const [selectedDebtor, setSelectedDebtor] =
-    useState<TableExtraordinaryInstallmentProps>({});
-
-  const [loading, setLoading] = useState(true);
-  const [isOpenModalDelete, setIsOpenModalDelete] = useState(false);
-  const [isOpenModalEdit, setIsOpenModalEdit] = useState(false);
-
   const isMobile = useMediaQuery("(max-width:880px)");
-
   const visbleHeaders = isMobile
     ? headers.filter((header) => rowsVisbleMobile.includes(header.key))
     : headers;
   const visbleActions = isMobile
     ? rowsActions.filter((action) => rowsVisbleMobile.includes(action.key))
     : rowsActions;
+
+  const [extraordinaryInstallments, setExtraordinaryInstallments] = useState<
+    TableExtraordinaryInstallmentProps[]
+  >([]);
+  const [selectedDebtor, setSelectedDebtor] =
+    useState<TableExtraordinaryInstallmentProps>({});
+  const [loading, setLoading] = useState(true);
+  const [isOpenModalDelete, setIsOpenModalDelete] = useState(false);
+  const [isOpenModalEdit, setIsOpenModalEdit] = useState(false);
+  const paginationProps = usePagination(extraordinaryInstallments);
+  const { addFlag } = useFlag();
+
+  const itemIdentifiersForUpdate: IExtraordinaryInstallments = {
+    creditProductCode: prospectData?.creditProducts[0].creditProductCode || "",
+    extraordinaryInstallments:
+      prospectData?.creditProducts[0]?.extraordinaryInstallments
+        ?.filter((installment) => {
+          const expectedId = `${prospectData?.creditProducts[0].creditProductCode},${installment.installmentDate},${installment.paymentChannelAbbreviatedName}`;
+          return expectedId === selectedDebtor?.id;
+        })
+        ?.map((installment) => ({
+          installmentDate:
+            typeof installment.installmentDate === "string"
+              ? installment.installmentDate
+              : new Date(installment.installmentDate).toISOString(),
+          installmentAmount: Number(installment.installmentAmount),
+          paymentChannelAbbreviatedName: String(
+            installment.paymentChannelAbbreviatedName,
+          ),
+        })) || [],
+    prospectId: prospectData?.prospectId || "",
+  };
 
   useEffect(() => {
     if (prospectData?.creditProducts) {
@@ -147,6 +170,13 @@ export const TableExtraordinaryInstallment = (
     setLoading(false);
   }, [prospectData, refreshKey]);
 
+  useEffect(() => {
+    if (extraordinary && Array.isArray(extraordinary)) {
+      setExtraordinaryInstallments(extraordinary);
+      setLoading(false);
+    }
+  }, [extraordinary]);
+
   const handleUpdateData = async (
     updatedDebtor: TableExtraordinaryInstallmentProps,
   ) => {
@@ -166,12 +196,39 @@ export const TableExtraordinaryInstallment = (
     }
   };
 
-  useEffect(() => {
-    if (extraordinary && Array.isArray(extraordinary)) {
-      setExtraordinaryInstallments(extraordinary);
-      setLoading(false);
+  const handleDeleteAction = async () => {
+    if (handleDelete && selectedDebtor.id) {
+      handleDelete(selectedDebtor.id as string);
+      setIsOpenModalDelete(false);
+    } else if (service) {
+      try {
+        await removeExtraordinaryInstallment(
+          businessUnitPublicCode ?? "",
+          itemIdentifiersForUpdate,
+        );
+
+        setSentData?.(itemIdentifiersForUpdate);
+        setIsOpenModalDelete(false);
+        handleClose?.();
+      } catch (error: unknown) {
+        const err = error as {
+          message?: string;
+          status?: number;
+          data?: { description?: string; code?: string };
+        };
+        const code = err?.data?.code ? `[${err.data.code}] ` : "";
+        const description =
+          code + (err?.message || "") + (err?.data?.description || "");
+
+        addFlag({
+          title: TextLabels.titleError,
+          description,
+          appearance: "danger",
+          duration: 5000,
+        });
+      }
     }
-  }, [extraordinary]);
+  };
 
   return (
     <TableExtraordinaryInstallmentUI
@@ -188,12 +245,14 @@ export const TableExtraordinaryInstallment = (
       setIsOpenModalDelete={setIsOpenModalDelete}
       setIsOpenModalEdit={setIsOpenModalEdit}
       handleUpdate={handleUpdateData}
-      usePagination={usePagination}
+      usePagination={paginationProps}
       setSentData={setSentData ?? (() => {})}
       handleClose={handleClose}
       setSelectedDebtor={setSelectedDebtor}
       handleDelete={handleDelete}
       service={service}
+      itemIdentifiersForUpdate={itemIdentifiersForUpdate}
+      handleDeleteAction={handleDeleteAction}
     />
   );
 };
